requirements_docs.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bandit==1.7.5
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
duckdb==0.8.0
entrypoints==0.4
Faker==18.9.0
fastapi==0.95.2
favicon==0.7.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
htbuilder==0.6.1
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.177
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.7
openapi==1.1.0
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pbr==5.11.1
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pluggy==1.0.0
posthog==3.0.1
protobuf==3.20.3
pyarrow==12.0.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.15.1
pymdown-extensions==10.0.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.9.0
pyreadline3==3.4.1
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
st-annotated-text==4.0.0
starlette==0.27.0
stevedore==5.1.0
coordinates
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
uvicorn==0.22.0
validators==0.20.0
Wand==0.6.11
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0